requirements.txt
Babel==1.3
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.3
alabaster==0.7.2
argh==0.26.1
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
docutils==0.12
livereload==2.3.2
pathtools==0.1.2
pycrypto==2.6.1
pytz==2014.10
six==1.9.0
snowballstemmer==1.2.0
sphinx-autobuild==0.5.0
sphinx-rtd-theme==0.1.7
tornado==4.1
watchdog==0.8.3
wsgiref==0.1.2
